Add tests for user index
